+++ addOns/network/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Allow to log client side HTTP traffic for debug purposes, using the name `org.zaproxy.addon.network.http`.
+
 ### Fixed
 - Do not pass-through requests to the local proxies themselves (e.g. ZAP domain, aliases).
 - Correctly handle concurrent requests (Issue 7838).